# cmd/gateway/openapi.yaml
openapi: 3.0.3
info:
  title: AI Gateway
  version: 1.7.0
  description: |
    Unified AI Gateway — OpenAI / Anthropic / Responses under one domain.
    Gateway auth: Bearer sk-gw-* or admin session JWT on /v1/*.
    Admin: Bearer JWT on /api/*. Errors: {error:{message,type}}.
servers:
  - url: http://localhost:8080
    description: local
  - url: https://ai.example.com
    description: production
paths:
...
  /v1/chat/completions:
    post:
      summary: Chat completions (OpenAI compat, gateway auth)
      tags: [gateway]
      security: [{ gatewayKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages]
              properties:
                model: { type: string, example: openai/gpt-4o-mini, maxLength: 256, description: "Prefer provider/model. Short ids still resolve." }
                messages: { type: array, items: { type: object } }
                stream: { type: boolean, default: false }
                temperature: { type: number }
      responses:
        '200': { description: completion }
        '401': { description: authentication_error }
        '429': { description: rate_limit_error }
  /v1/completions:
    post:
      summary: Completions (OpenAI compat)
      tags: [gateway]
      security: [{ gatewayKey: [] }]
      responses:
        '200': { description: completion }
  /v1/embeddings:
    post:
      summary: Embeddings
      tags: [gateway]
      security: [{ gatewayKey: [] }]
      responses:
        '200': { description: embeddings }
  /v1/messages:
    post:
      summary: Anthropic Messages compat
      tags: [gateway]
      security: [{ gatewayKey: [] }]
      responses:
        '200': { description: message }
  /v1/responses:
    post:
      summary: OpenAI Responses API (native or translated)
      tags: [gateway]
      security: [{ gatewayKey: [] }]
      responses:
        '200': { description: response }
  /v1/models:
    get:
      summary: Models with id=provider/model (aliases remain bare)
      tags: [gateway]
      security: [{ gatewayKey: [] }]
      responses:
        '200':
          description: models
          headers:
            X-Cache: { schema: { type: string, enum: [HIT, MISS] } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { type: string, example: list }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, example: openai/gpt-4o-mini }
                        model_id: { type: string, example: gpt-4o-mini }
                        owned_by: { type: string, example: openai }
...
  /api/auth/passkey/login/begin:
    post:
      summary: Begin passkey login (public, brute-force rate limited)
      tags: [auth]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
      responses:
        '200': { description: WebAuthn assertion options + session id }
        '404': { description: user not found or no credentials }
        '429': { description: too many attempts }
  /api/auth/passkey/login/finish:
    post:
      summary: Finish passkey login (public) — returns admin JWT + session cookie
      tags: [auth]
      parameters:
        - name: X-WebAuthn-Session
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WebAuthn credential (assertion response); session may also arrive via body.session
      responses:
        '200':
          description: token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  username: { type: string }
                  role: { type: string }
        '401': { description: verification failed }
        '400': { description: missing/unknown session }
  /api/auth/recovery/verify:
    post:
      summary: Verify a recovery code and mint a new session (public, brute-force rate limited; code is single-use)
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username: { type: string }
                recovery_code: { type: string, description: alias — code }
                code: { type: string }
      responses:
        '200':
          description: new admin session
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  username: { type: string }
                  role: { type: string }
        '400': { description: username and code required }
        '401': { description: invalid recovery code }
        '429': { description: too many attempts }
  /api/auth/passkey/register/begin:
    post:
      summary: Begin passkey registration for the caller (admin may pass ?user_id for another account)
      tags: [auth]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: query
          required: false
          schema: { type: string }
      responses:
        '200': { description: WebAuthn credential creation options + session id }
        '401': { description: unauthorized }
        '403': { description: forbidden (non-admin targeting another user) }
  /api/auth/passkey/register/finish:
    post:
      summary: Finish passkey registration (attestation; session via X-WebAuthn-Session header/query/body)
      tags: [auth]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: WebAuthn credential (attestation response)
      responses:
        '200': { description: credential stored }
        '400': { description: missing/unknown session }
        '401': { description: unauthorized }
  /api/auth/passkey/credentials:
    get:
      summary: List the caller's (or ?user_id=…) registered passkeys
      tags: [auth]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: query
          required: false
          schema: { type: string }
      responses:
        '200': { description: credential list }
  /api/auth/passkey/recovery/generate:
    post:
      summary: Generate/rotate recovery codes for the caller (admin may target ?user_id)
      tags: [auth]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: query
          required: false
          schema: { type: string }
      responses:
        '200': { description: recovery codes shown once }
        '403': { description: forbidden }
  /api/auth/passkey/disable:
    post:
      summary: Disable passkey login for the caller (admin may target ?user_id)
      tags: [auth]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: query
          required: false
          schema: { type: string }
      responses:
        '200': { description: disabled }
        '403': { description: forbidden }
...
  /api/provider-models:
    get:
      summary: List discovered/manual provider models
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: provider models }
    post:
      summary: Manually add a provider model (audited, admin/member)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_id, model_id]
              properties:
                provider_id: { type: string, maxLength: 64 }
                model_id: { type: string, maxLength: 256 }
                display_name: { type: string }
                context_window: { type: integer }
                max_output: { type: integer }
                input_cost: { type: number }
                output_cost: { type: number }
                reasoning: { type: boolean }
                tool_call: { type: boolean }
                attachment: { type: boolean }
      responses:
        '201': { description: created }
        '400': { description: provider_id and model_id required }
  /api/provider-models/{id}:
    put:
      summary: Update a manual provider model (admin only, audited)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: ProviderModel fields to persist
      responses:
        '200': { description: ok }
        '400': { description: invalid json }
        '500': { description: update failed }
    delete:
      summary: Delete a provider model (admin only, audited)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: deleted }
  /api/provider-models/{id}/enrich:
    post:
      summary: Enrich a provider model from the models.dev catalog (admin only)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: enriched }
        '500': { description: enrichment failed }
  /api/providers/{id}/discover:
    post:
      summary: Run model discovery against one provider's /models endpoint (admin only, audited)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: discovered model ids
          content:
            application/json:
              schema:
                type: object
                properties:
                  discovered: { type: array, items: { type: string } }
        '404': { description: provider not found }
  /api/discover-all:
    post:
      summary: Run discovery across all providers (admin only, audited)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: discovery report }
...